chore!: delete unnecessary commands
