+++ commands/discord_commands.yaml
@@ -5,17 +5,6 @@
       description: The message to chat with the bot.
       type: 3 # string
       required: true
-  
-- name: course
-  description: Get information about an UML course
-  options:
-    - name: question
-      description: Ask a question about a course
-      type: 3
-      required: true
 
 - name: dog
   description: Fetches a random dog image.
-
-- name: reset
-  description: Resets the rate limits for every user on the server
