+++ commands/register_commands.py
@@ -8,13 +8,33 @@ load_dotenv()
 TOKEN = os.getenv('TOKEN')
 APPLICATION_ID = os.getenv('ID')
 URL = f"https://discord.com/api/v9/applications/{APPLICATION_ID}/commands"
+headers = {"Authorization": f"Bot {TOKEN}", "Content-Type": "application/json"}
+
+def delete():
+    dict = requests.get(URL, headers=headers)
+    registered_commands = dict.json()
+    registered_names = {command["name"] for command in registered_commands}
+    registered_table = {command["name"]: command["id"] for command in registered_commands}
+
+    with open("discord_commands.yaml", "r") as file:
+        yaml_content = file.read()
+
+    commands = yaml.safe_load(yaml_content)
+    local_names = {command.get("name") for command in commands}
+
+    delete_commands = registered_names - local_names
+
+    for command in delete_commands:
+        id = registered_table[command]
+        requests.delete(f"{URL}/{id}", headers=headers)
+        print(f"Deleted {command}")
+        time.sleep(2)
 
 def register():
     with open("discord_commands.yaml", "r") as file:
         yaml_content = file.read()
 
     commands = yaml.safe_load(yaml_content)
-    headers = {"Authorization": f"Bot {TOKEN}", "Content-Type": "application/json"}
 
     # Send the POST request for each command
     for command in commands:
@@ -23,4 +43,6 @@ def register():
         print(f"Command {command_name} created: {response.status_code}")
         time.sleep(3)
 
-register()
+if __name__ == "__main__":
+    delete()
+    register()
